Intel_Operation_Sqli/src/requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0  # ItsDangerous est souvent utilisé pour signer et crypter des données sensibles, généralement dans le contexte des sessions web ou des jetons d'authentification.
Jinja2==2.11.2
MarkupSafe==1.1.1    #  MarkupSafe est utilisé pour échapper des données avant de les inclure dans un modèle Jinja2 afin d'éviter les attaques XSS (Cross-Site Scripting).
mysql==0.0.2
mysql-connector-python==8.0.21
mysqlclient==2.0.1
protobuf==3.13.0
python-dotenv==0.14.0
six==1.15.0
Werkzeug==1.0.1
